feat: move classes into different files
