+++ aries_askar_upgrade/error.py
@@ -1,0 +1,4 @@
+
+
+class UpgradeError(Exception):
+    pass
